vk: studio: extract all model/cache stuff into separate files
